removed code from TradeWidget
1. moved into sub folders
